+++ packages/_private/browser-baseline/src/index.js
@@ -1,0 +1,23 @@
+/**
+ * @typedef {keyof typeof baselines} BaselineKeys
+ */
+import * as baselines from './baselines/index.js';
+
+/**
+ * Get the baseline configuration for a specific version
+ * @param {BaselineKeys} version The version to get the baseline for
+ * @returns {string[]} The baseline configuration for the specified version
+ * @throws Will throw an error if the specified version does not exist
+ */
+export const getBaseline = (version) => {
+  if (!baselines[version]) {
+    throw new Error(`Baseline for version ${version} not found. Available versions: ${Object.keys(baselines).join(', ')}`);
+  }
+  return baselines[version];
+};
+
+/**
+ * Get a list of available baseline versions
+ * @returns {string[]} An array of available baseline versions
+ */
+export const getAvailableBaselines = () => Object.keys(baselines);